# public/openapi.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Empresa de Transportes El Cumbe - API RESTful",
                 "version":  "1.0.0",
                 "description":  "EspecificaciÃ³n oficial de servicios RESTful y API MÃ³vil para la plataforma de transporte interprovincial El Cumbe. Incluye servicios de autenticaciÃ³n JWT, bÃºsqueda de salidas, reserva de asientos, pasarela de pagos Culqi y gestiÃ³n de encomiendas."
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3000",
                        "description":  "Servidor Local de Desarrollo"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Ingrese el token JWT obtenido en la API de Login MÃ³vil (/api/movil/login)"
                                                              }
                                           },
...
                   },
    "paths":  {
...
                  "/api/movil/conductor/viajes":  {
                                                      "get":  {
                                                                  "tags":  [
                                                                               "Conductor"
                                                                           ],
                                                                  "summary":  "Listar viajes del conductor autenticado",
                                                                  "security":  [
                                                                                   {
                                                                                       "bearerAuth":  [

                                                                                                      ]
                                                                                   }
                                                                               ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Operación exitosa"
                                                                                            },
                                                                                    "400":  {
                                                                                                "description":  "Solicitud inválida"
                                                                                            },
                                                                                    "401":  {
                                                                                                "description":  "Token ausente o inválido"
                                                                                            },
                                                                                    "403":  {
                                                                                                "description":  "Rol o recurso no autorizado"
                                                                                            },
                                                                                    "500":  {
                                                                                                "description":  "Error interno"
                                                                                            }
                                                                                }
                                                              },
                                                      "put":  {
                                                                  "tags":  [
                                                                               "Conductor"
                                                                           ],
                                                                  "summary":  "Actualizar un viaje propio",
                                                                  "security":  [
                                                                                   {
                                                                                       "bearerAuth":  [

                                                                                                      ]
                                                                                   }
                                                                               ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Operación exitosa"
                                                                                            },
                                                                                    "400":  {
                                                                                                "description":  "Solicitud inválida"
                                                                                            },
                                                                                    "401":  {
                                                                                                "description":  "Token ausente o inválido"
                                                                                            },
                                                                                    "403":  {
                                                                                                "description":  "Rol o recurso no autorizado"
                                                                                            },
                                                                                    "500":  {
                                                                                                "description":  "Error interno"
                                                                                            }
                                                                                },
                                                                  "requestBody":  {
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "type":  "object"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  },
                                                                                      "required":  true
                                                                                  }
                                                              }
                                                  },
...
                  "/api/movil/operario/pasajeros":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Operario"
                                                                             ],
                                                                    "summary":  "Consultar manifiesto de pasajeros",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Operación exitosa"
                                                                                              },
                                                                                      "400":  {
                                                                                                  "description":  "Solicitud inválida"
                                                                                              },
                                                                                      "401":  {
                                                                                                  "description":  "Token ausente o inválido"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Rol o recurso no autorizado"
                                                                                              },
                                                                                      "500":  {
                                                                                                  "description":  "Error interno"
                                                                                              }
                                                                                  }
                                                                },
                                                        "put":  {
                                                                    "tags":  [
                                                                                 "Operario"
                                                                             ],
                                                                    "summary":  "Actualizar estado de abordaje",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Operación exitosa"
                                                                                              },
                                                                                      "400":  {
                                                                                                  "description":  "Solicitud inválida"
                                                                                              },
                                                                                      "401":  {
                                                                                                  "description":  "Token ausente o inválido"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Rol o recurso no autorizado"
                                                                                              },
                                                                                      "500":  {
                                                                                                  "description":  "Error interno"
                                                                                              }
                                                                                  },
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "object"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    }
                                                                }
                                                    },
...
              }
}